separating Card with AttachmentTab
